ll: move `sort`
